Add HeightViewModel

+++ core/model/src/main/java/com/dante/calorietracker/core/model/UserInfo.kt
@@ -4,7 +4,7 @@ data class UserInfo(
     val gender: Gender,
     val age: Int,
     val weight: Float,
-    val height: Float,
+    val height: Int,
     val activityLevel: ActivityLevel,
     val goalType: GoalType,
     val carbRatio: Float,
